Fix window/buffer restoration on session open
When loading the session, the current buffer number was noted before
sourcing the session file, and then wiped out afterwards, but this can
wipe out the last buffer loaded by the session in vim 8.1, because files
seem to be loaded into an existing buffer rather than into a new buffer,
so the session loads a file into a buffer which then gets wiped out :-(

Fix by just avoiding deleting the code to note the current buffer before
loading the session and wipe it out afterwards. Seems like this might
have been unnecessary for a long time anyway, I've tested with vim 7.4
and 8.1, and it works fine with sessions saved and restored in both.