Sort sessions by modified date, most recent first
This makes sense for my config, because my mapping to load the session
list first saves the current session, which means the last modified
timestamps on sessions are effectively most recently used timestamps.

I think this makes a lot of sense in general, so I'll come back and look
at updating this fork so this works by default, with no custom mappings.